Move files to `select/` directory
