+++ packages/labs/src/blueprint-labs.scss
@@ -5,6 +5,6 @@
  * and https://github.com/palantir/blueprint/blob/master/PATENTS
  */
 
-@import "./multi-select/multi-select";
+@import "./select/multi-select";
 @import "./select/select";
 @import "./tag-input/tag-input";
