+++ packages/labs/src/select/_multi-select.scss
